requirements.txt
pylint==1.8.1
Pillow==4.3.0
tqdm==4.19.5
